Resetting all settings from within the UI menu. (#436)
Able to reset the customer setting from the menu. It will bring a waring dialog to ensure to reset by mistake.  So no need to add a file to sd-card to reset the settings.
Ignore lcd mode only if you reset all settings and to show the info dialog.
Otherwise, it won't show if you currently in CLEAN_MODE_SWITCHING_SUPPORT.