Fixed tool switching for multiple extruders, change the default step for Baby step
Status screen fix.
When using multiple extruders, the tempreture display checks the number of extruders instead of the number of tools. Because of this
the temperature on the screen cycles between non-existing tools.

Baby step
the default value for the change is 0.1mm.
This is value is too big for Baby step.
Changing it to 0.01mm makes more sense.