NEW Hybrid Red Menu Icon PACK
A brand new Menu Icon system. It was named Hybrid Red Menu because it contains icons from multiple areas. It has MKS TFT icon. It has BTT TFT icon inside. Most of the icons are dominated by red. That is why it became the name of Hybrid Red Menu. It works accurately and hopefully flawlessly. Feel free to contact me with questions.

She's in a ZIP folder because I couldn't create a folder for her in the root directory. And it would look better that way.

Tested on: BIGTREETECH TFT35 display

Hello, AntoszHUN