The latest patch is Hybrid Red Menu icon
I fixed the missing icons for all display types. I hope all icons are correct. The names Reset and reset_value are not clear for the different types. I didn't poke these because one display uses the name Reset and the other uses reset_value? It's a little confusing to me. But I didn't vary it. If you find a mistake, report it!

Hello, AntoszHUN