Option for Mixing_extruder
If the Mixing_extruder option is selected in Marlin, the printer uses two extruders, but reports only one extruder at the M115 command.

The FW display overwrites the manually entered number of extruders with what the printer reports.
This option disables the value override, and the value specified in Configuration.h is used

resolves # 1055