Reduced path to start ABL for Delta printers
With Delta printers there is no need for a second layer/page for ABL because on second layer/page there is only one icon to start ABL.

This PR reduce path to start ABL only for Delta printers, now ABL is starting when clicking on ABL button for ease of use. Before, you had to click on ABL and then on Start on the 2nd layer/page.