Puristic changes and updating of the translation
I wanted to unify systematics of SD and USB descriptions: SD card and USB disk, TFT SD and TFT USB, the same sentence syntax, added spaces. From now on, no one will mistake TFT USB for the USB of the motherboard, because it is clear.
The translation takes advantage of this.