Update language_fr.h
French customization to fit with max size of 10 char (label_clear), adapt the wording correctly ( label_tft_sd_inserted & label_tft_sd_removed) and adapt to comprehensive button (label_emergencystop) ( "Arrêt" in french is similar to Stop, but not not urgency. The correct spelling will be "Arrêt d'urgence" or "Stop d'urgence" but the size is limited to 10 char. So "Urgence" is more comprehensive with the bmp button with the stop)