Added MKS TFT32L V3.0 and fixed horizontal mirror display.
For this display, the SS and GS bits must be identical.
Not having another MKS 28 or 32 display, I added its management in addition to not break the existing one.
For information, the MKS 32 series shares the firmware of the MKS 28 series. the firmware must be named "MKSTFT28" to be taken into account by the boot loader, so naming its folder TFT32 is counterproductive and is therefore named "TFT28 " like the firmware.