fixes the emulator mode which is active even on non-compatible displays
fixes the emulator mode which is active even on non-compatible displays and especially without the "#define ST7920_EMULATOR" directive. a long press still offers the change of mode.

the origin of the bug is in the "variants.h" file.
"#define HAS_EMULATOR (defined(ST7920_EMULATOR) || defined(LCD2004_EMULATOR))"
in this case, "HAS_EMULATOR" is loaded with a false or true value.
however in the code, it is its existence that is tested by "#ifdef HAS_EMULATOR"
changing this code to this defines it or not and fixes the bug.
#if (defined(ST7920_EMULATOR) || defined(LCD2004_EMULATOR))
   #define HAS_EMULATOR
#endif

added a test in "Printing.c" because "menuMarlinMode" does not exist if "HAS_EMULATOR" does not exist either. (test that exists elsewhere online 909).

now a long press no longer proposes the change of mode.